+++ test/replique/compilement/ns_mappings_cljs_test.cljs
@@ -1,7 +1,8 @@
 (ns replique.compliment.ns-mappings-cljs-test
+  (:require [clojure.string :as st :refer [split] #_:rename #_{split str-split}])
   (:import [goog.events EventType])
   (:require-macros [replique.compliment.ns-mappings-clj-test
-                    :as cljs-ns-m :refer [my-macro]]))
+                    :as cljs-ns-m :refer [my-macro] #_:rename #_{my-macro my-macro-2}]))
 
 (defn my-fn "rr" [e f & {:keys [e r]}])
 (def gg nil)